// pc/src/app/types/common.type.ts
export interface ResponseBody {
    data: any;
    msg: string;
    status: number;
}

interface ListExtend {
    noread?: number;
    isread?: number;
    allread?: number;
    totalBetAmount?: string;
    totalValidBetAmount?: string;
    totalPayout?: string;
    totalNetAmount?: string;
    currBetAmount?: string;
    currValidBetAmount?: string;
    currPayout?: string;
    currNetAmount?: string;
}

export interface ResponseList {
    list: Array<any>;
    pageNo: number;
    pageSize: number;
    total: number;
    totalPage: number;
    extendBean?: ListExtend;
}
// ...
export interface BetPlatList {
    type: number;
    desc: string;
    bets: Array<{ gameName: string; gameCode: string; gameImg: string; selected?: boolean }>;
}

export interface SelectPlat {
    type: number;
    gameName: string;
    gameCode: string;
}